requirement.txt
numpy==1.18.5
pyyaml==5.3.1
tqdm==4.47.0
scikit-learn==0.23.1
nltk==3.5
paddlepaddle-gpu==2.2.0.post112
paddlenlp==2.0.7
torch==1.7.1
transformers==4.15.0
pytorch_transformers==1.2.0
easydict
